requirements.txt
fastapi~=0.117.1
uvicorn~=0.37.0
starlette~=0.48.0
pydantic==2.11.9
annotated-types~=0.7.0
python-dateutil~=2.8.2
numpy~=1.24.0
pandas~=1.5.3
scikit-learn==1.5.1
joblib~=1.5.2
PyYAML~=6.0
pytz~=2022.7.1
click~=8.3.0
Jinja2~=3.1.2
MarkupSafe~=2.1.2
six~=1.16.0
websockets~=15.0.1
httptools~=0.6.4
watchfiles~=1.1.0
pymongo
passlib==1.7.4
python-jose[cryptography]
python-multipart
openpyxl
email-validator
python-dotenv
bcrypt==3.2.0
gunicorn
motor